refactor(trading): separate middleware loading data to thunk
